requirements.txt
flask==2.3.3
requests_oauthlib==1.3.1
transformers==4.35.2
python-dotenv==1.0.0
pandas==2.1.3
torch==2.1.1
requests==2.31.0
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
numpy==1.26.2
tqdm==4.66.1
regex==2023.10.3
packaging==23.2
filelock==3.13.1
certifi==2023.11.17
urllib3==2.1.0
idna==3.6
charset-normalizer==3.3.2 